Tests: Simplified and renamed spatial exch
